# Use AtomicU64 instead of u64 for the position field in Receiver