--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "async-broadcast"
-version = "0.7.1"
+version = "0.8.0"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/smol-rs/async-broadcast"
 documentation = "https://docs.rs/async-broadcast"