The `recv()`, `recv_direct()`, `recv_blocking()`, and `try_recv()` methods currently require `&mut self` to modify the value of the `position` field. By using AtomicU64 for the `position` field eliminates the need for mutability.

Regarding performance, it's worth noting that on most processor architectures, there is no much difference between an atomic store and a regular non-atomic store (source: [1](https://marabos.nl/atomics/memory-ordering.html#the-memory-model), [2](https://marabos.nl/atomics/hardware.html#load-store-ops)). 

Additionally, Atomic operations are generally not as expensive as mutexes, as mutexes themselves are implemented using atomic operations but with additional overhead for locking.

**This PR has breaking API changes.**

Fixes issue #66