@zeenix 2024-07-04
> Regarding performance, it's worth noting that on most processor architectures, there is no much difference between an atomic store and a regular non-atomic store ([source](https://marabos.nl/atomics/memory-ordering.html#the-memory-model)).

Thanks for the citing a source on this. I'll read it later..

> Additionally, Atomic operations are generally not as expensive as mutexes, as mutexes themselves are implemented using atomic operations but with additional overhead for locking.

It would be nice to have some numbers on this but I guess we know for sure that at least atomics are not **more** expensive. So if atomics are not more expensive than their non-atomic primitive types for most machines out there, there is no downside to using atomics and that's good enough for me.